application/backend/requirements-test.txt
pytest
pytest_postgresql
pytest-flask
pytest-mock
pymaybe
psycopg
